provided vital information
necessary for fundamental understanding by those that have become beseechers of unfathomable knowledge; who walk upon the path of righteousness bound solely in the remains of woven fabric and silken strands of absolute and holy truth righteousness and understanding. To lead the way for those tempered to beseech understanding and the ability and duty to smite those that are none other than unholy sycophants possessing none but the sole purpose to siphon one's faith and conviction, one's knowledge through understanding, and above all, the eventual cold-hearted truth, by eradicating all traces and remnants of one's fleeting innocence when devouring all that remains of one's shattered and feeble soul. Effectively eliminating any indication of one's identification, recognition, or acknowledgment of what once had been deemed "individual" or singular representation of the self, the id, or ego, never to be heard from again.